Create README.md
The scripts in this repo are created with the purpose of modelling the factors associated with critical and commercial success of films released in the US in 2015. The commercial success aspect is measured by the Box Office numbers for all films released in 2015 - including their revenue in 2016 if they were released toward the end of the calendar year. Critical success is measured by the number of Oscar nominations and wins, and number of nominations and wins for awards given by five critics' associations in the US.